Fixed pipeline resolve and centralized ensemble builder
