+++ evalml/automl/automl_algorithm/default_algorithm.py
@@ -18,10 +18,7 @@ from evalml.pipelines.components.utils import (
     get_estimators,
     handle_component_class,
 )
-from evalml.pipelines.utils import (
-    _make_stacked_ensemble_pipeline,
-    make_pipeline,
-)
+from evalml.pipelines.utils import make_pipeline
 from evalml.problem_types import is_regression
 
 
@@ -220,78 +217,6 @@ class DefaultAlgorithm(AutoMLAlgorithm):
         for pipeline in pipelines:
             self._create_tuner(pipeline)
         return pipelines
-
-    def _create_ensemble(self):
-        next_batch = []
-
-        # Custom Stacked Pipelines
-        ensembler_component_graph = {}
-        final_components = []
-        problem_type = None
-        n_jobs_ensemble = 1 if self.text_in_ensembling else self.n_jobs
-
-        for model_type, best_info in self._best_pipeline_info.items():
-
-            def _make_new_component_name(component_name):
-                return str(model_type) + " Pipeline - " + component_name
-
-            pipeline = best_info["pipeline"]
-            if problem_type is None:
-                problem_type = pipeline.problem_type
-            final_component = None
-            ensemble_y = "y"
-            for name, component_list in pipeline.component_graph.component_dict.items():
-                new_component_list = []
-                new_component_name = _make_new_component_name(name)
-                for i, item in enumerate(component_list):
-                    if i == 0:
-                        fitted_comp = pipeline.component_graph[item.name]
-                        new_component_list.append(fitted_comp)
-                    elif isinstance(item, str) and item not in ["X", "y"]:
-                        new_component_list.append(_make_new_component_name(item))
-                    else:
-                        new_component_list.append(item)
-                    if i != 0 and item.endswith(".y"):
-                        ensemble_y = _make_new_component_name(item)
-                ensembler_component_graph[new_component_name] = new_component_list
-                final_component = new_component_name
-            final_components.append(final_component)
-
-        ensemble = _make_stacked_ensemble_pipeline(
-            problem_type,
-            component_graph=ensembler_component_graph,
-            final_components=final_components,
-            random_seed=self.random_seed,
-            n_jobs=n_jobs_ensemble,
-            ensemble_y=ensemble_y,
-        )
-        next_batch.append(ensemble)
-
-        # Sklearn Stacked Pipelines
-        input_pipelines = []
-        for pipeline_dict in self._best_pipeline_info.values():
-            pipeline = pipeline_dict["pipeline"]
-            pipeline_params = pipeline_dict["parameters"]
-            parameters = self._transform_parameters(pipeline, pipeline_params)
-            if (
-                "Select Columns Transformer"
-                in pipeline.component_graph.component_instances
-            ):
-                parameters.update(
-                    {"Select Columns Transformer": {"columns": self._selected_cols}}
-                )
-            input_pipelines.append(
-                pipeline.new(parameters=parameters, random_seed=self.random_seed)
-            )
-        ensemble = _make_stacked_ensemble_pipeline(
-            problem_type,
-            input_pipelines=input_pipelines,
-            random_seed=self.random_seed,
-            n_jobs=n_jobs_ensemble,
-            use_sklearn=True,
-        )
-        next_batch.append(ensemble)
-        return next_batch
 
     def _create_n_pipelines(self, pipelines, n):
         next_batch = []
